back-end-no-ids/code/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==2.6.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.3.3
libclang==18.1.1
lightgbm==4.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyJWT==2.8.0
PyPDF2==3.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.14.1
